refactor(artifacts): rename _graphql_fragments.py -> _gqlutils.py
